fixup! secboot: address review comments
