+++ cmd/snap-bootstrap/main.go
@@ -41,7 +41,7 @@ such as initramfs.
 )
 
 func main() {
-	secboot.HijackAndRunArgon2OutOfProcessHandlerOnArg([]string{"--argon2-proc"})
+	secboot.HijackAndRunArgon2OutOfProcessHandlerOnArg([]string{"argon2-proc"})
 
 	err := run(os.Args[1:])
 	if err != nil {
