+++ secboot/argon2_out_of_process.go
@@ -1,0 +1,36 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2025 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package secboot
+
+import "os"
+
+func isOutOfProcessArgon2KDFMode(args []string) bool {
+	if len(os.Args) != len(args)+1 {
+		return false
+	}
+
+	for i := 0; i < len(args); i++ {
+		if os.Args[i+1] != args[i] {
+			return false
+		}
+	}
+
+	return true
+}
